修复删除图书bug

+++ src/main/java/com/example/library/service/BookService.java
@@ -91,7 +91,7 @@ public class BookService {
      * @param id 主键
      * @return true or false
      */
-    public void deleteBook(Integer id) {
+    public void deleteBook(Long id) {
         bookMapper.deleteById(id);
     }
